+++ src/app/app.module.ts
@@ -89,6 +89,7 @@ import { DateComponent } from './lib/widgets/date/date.component';
 import { TextAreaComponent } from './lib/widgets/textarea/textarea.component';
 import {DatetimeComponent} from './lib/widgets/datetime/datetime.component';
 import {LabelComponent} from './lib/widgets/label/label.component';
+import { EditableLinkIdComponent } from './lib/widgets/editable-link-id/editable-link-id.component';
 import {CdkCopyToClipboard} from "@angular/cdk/clipboard";
 import {CodemirrorModule} from "@ctrl/ngx-codemirror";
 import {HelpTextComponent} from "./lib/widgets/help-text/help-text.component";
@@ -156,6 +157,7 @@ import {HelpTextComponent} from "./lib/widgets/help-text/help-text.component";
     DateComponent,
     TextAreaComponent,
     DatetimeComponent,
+    EditableLinkIdComponent,
     HelpTextComponent
   ],
   imports: [