+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [9.6.10] 2024-11-01
+### Added
+- Added functionality for item validation and corresponding error display.
+- Added editable link id field.
+
 ## [9.6.9] 2024-10-17
 ### Changed
 - Added Group Item Control options.